Let a block be as big as the selection, not 8x8
MAX_BLOCK = 8 was invented when blocks were added — a block is only
references into the tile bank, so neither the hardware nor the format
asked for it, and naming a larger selection was refused for no reason.

What does constrain a block is the generated C, in two places:
_DrawBlock takes width and height as u8, and it took `base` — the byte
offset into the _Blocks table — as u8 too. That offset is the one that
would really have broken, since a couple of large blocks push it past
255. It is now u16, which is what makes big blocks safe rather than
merely permitted, and the axis cap is 255.

Past 32x24 a block cannot be stamped onto the screen in one call, but
the data is still valid, so that is left to judgement rather than
enforced. The "+ Block" size pickers were selects enumerating the cap,
which at 255 entries is no use; they are number inputs now.

Verified by re-exporting the demo's tileset with the widened helper and
running the ROM — the door it stamps is on that path.

Also carries the demo's in-app intro edits.

+++ demo_project/content/tiles.h
@@ -652,7 +652,7 @@ const unsigned char g_Tiles_Blocks[] =
 //
 // Example:
 //   g_Tiles_DrawBlock(10, 5, G_TILES_COIN_SPIN_BASE, G_TILES_COIN_SPIN_W, G_TILES_COIN_SPIN_H);
-static void g_Tiles_DrawBlock(u8 x, u8 y, u8 base, u8 w, u8 h)
+static void g_Tiles_DrawBlock(u8 x, u8 y, u16 base, u8 w, u8 h)
 {
 	VDP_WriteLayout_GM2(g_Tiles_Blocks + base, x, y, w, h);
 }
